Location used to get weather info

+++ app/src/main/java/com/tnurdinov/dvtweather/api/Item.kt
@@ -9,7 +9,7 @@ data class Item(
     @SerializedName("clouds"     ) var clouds     : Clouds?            = Clouds(),
     @SerializedName("wind"       ) var wind       : Wind?              = Wind(),
     @SerializedName("visibility" ) var visibility : Int?               = null,
-    @SerializedName("pop"        ) var pop        : Int?               = null,
+    @SerializedName("pop"        ) var pop        : Double?            = null,
     @SerializedName("sys"        ) var sys        : Sys?               = Sys(),
     @SerializedName("dt_txt"     ) var dtTxt      : String?            = null
 )
